requirements.txt
# ==============================================================================
# PlantaeClassScanner Dependencies - PYTHON 3.11 VERIFIED
# Python 3.11+ | Ubuntu 22.04 | RasPi 4B (8GB)
# Last verified: 2025-10-10
# All versions confirmed available for Python 3.11
# ==============================================================================

# ============== CORE ML/CV (Version-locked for compatibility) ==============
tensorflow==2.15.0
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.4
scipy==1.12.0

# ============== DATA HANDLING ==============
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2

# ============== HARDWARE (RasPi - conditional install) ==============
RPi.GPIO==0.7.1; platform_machine=="aarch64" or platform_machine=="armv7l"
gpiozero==2.0.1; platform_machine=="aarch64" or platform_machine=="armv7l"

# ============== API/DATA COLLECTION ==============
requests==2.31.0
tqdm==4.66.1
beautifulsoup4==4.12.3

# ============== MODEL TRAINING UTILITIES ==============
scikit-learn==1.4.1.post1
imgaug==0.4.0
split-folders==0.5.1

# ============== VISUALIZATION ==============
plotly==5.18.0

# ============== DEPLOYMENT ==============
flask==3.0.1
waitress==2.1.2

# ============== DEVELOPMENT TOOLS ==============
pytest==7.4.4
black==24.1.1
pylint==3.0.3
ipython==8.20.0
jupyter==1.0.0

# ============== MODEL OPTIMIZATION ==============
onnx==1.15.0
onnxruntime==1.16.3

# ============== UTILITIES ==============
pyyaml==6.0.1
python-dotenv==1.0.1
colorama==0.4.6

# ============== ADDITIONAL DEPENDENCIES ==============
h5py==3.10.0
protobuf==4.23.4